pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "magick_mind"
version = "0.0.3"
authors = [
    { name = "Adrian", email = "minoak@globalmagicko.com" },
    { name = "Turtle", email = "turtle@globalmagicko.com" },
    { name = "Min Thu Wai", email = "minthu@globalmagicko.com" },
]
description = "A package for the Magick Mind project by General Magicko Industries"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "litellm>=1.60.8",
    "nano-vectordb>=0.0.4.3",
    "networkx>=3.4.2",
    "numpy>=2.2.2",
    "pipmaster>=0.4.0",
    "pydantic-settings>=2.7.1",
    "tenacity>=9.0.0",
    "pygments>=2.17.2",
    "bs4>=0.0.2",
]

[project.optional-dependencies]
semantic = ["lightrag-hku>=1.1.5"]

episodic = ["pymongo>=4.11.1"]

test = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.5"
]

[project.urls]
Homepage = "https://github.com/General-Magick-Industries/AGD_Magick_Mind"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["magick_mind"]
markers = [
    "asyncio: mark test as async test",
]

# [tool.hatch.build.targets.wheel]
# packages = ["libs"]

# [tool.hatch.build.targets.wheel.force-include]
# "libs" = "super_brain"
